Rename appender classes to GecLog4jAppender and GecLogbackAppender, respectively
