+++ java/gec-appender-log4j/src/main/java/com/greplin/gec/GecLog4jAppender.java
@@ -16,16 +16,12 @@
 
 package com.greplin.gec;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.classic.spi.IThrowableProxy;
-import ch.qos.logback.classic.spi.StackTraceElementProxy;
-import ch.qos.logback.classic.spi.ThrowableProxy;
-import ch.qos.logback.classic.spi.ThrowableProxyUtil;
-import ch.qos.logback.core.AppenderBase;
-import ch.qos.logback.core.CoreConstants;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
+import org.apache.commons.lang.exception.ExceptionUtils;
+import org.apache.log4j.AppenderSkeleton;
+import org.apache.log4j.Level;
+import org.apache.log4j.spi.LoggingEvent;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -38,10 +34,11 @@ import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicLong;
 
+
 /**
  * log4j appender that writes exceptions to a file to be picked up by upload.py.
  */
-public final class GecAppender extends AppenderBase<ILoggingEvent> {
+public final class GecLog4jAppender extends AppenderSkeleton {
   /**
    * Number of prefixes to use to reduce risk of server invocations overwriting
    * each other's error logs.
@@ -96,35 +93,22 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
   /**
    * Set of classes that only exist to contain exceptions.
    */
-  private final Set<String> passthroughExceptions;
+  private final Set<Class<? extends Throwable>> passthroughExceptions;
 
   /**
    * Creates a new appender.
    */
-  public GecAppender() {
-    // setThreshold(Level.ERROR); // FIXME: replaced by filters ?
-    this.passthroughExceptions = new HashSet<String>();
-    this.passthroughExceptions.add(
-        InvocationTargetException.class.getCanonicalName());
+  public GecLog4jAppender() {
+    setThreshold(Level.ERROR);
+    this.passthroughExceptions = new HashSet<Class<? extends Throwable>>();
+    this.passthroughExceptions.add(InvocationTargetException.class);
   }
 
   @Override
-  public void start() {
-    /* FIXME outdated docs ?
-    if (this.layout == null) {
-      addError("No layout set for the appender named [" + name + "].");
-      return;
-    }
-    */
-    super.start();
-  }
-
-  @Override
-  protected void append(final ILoggingEvent loggingEvent) {
-
+  protected void append(final LoggingEvent loggingEvent) {
     try {
-      if (loggingEvent.getThrowableProxy() == null
-          && loggingEvent.getLevel().toInt() < Level.ERROR_INT) {
+      if (loggingEvent.getThrowableInformation() == null
+          && loggingEvent.getLevel().toInt() < this.threshold.toInt()) {
         // Ignore non-exceptions below our threshold.
         return;
       }
@@ -134,15 +118,17 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
       File output = new File(this.outputDirectory, filename + ".writing");
       Writer writer = new FileWriter(output);
 
-      if (loggingEvent.getThrowableProxy() == null) {
+      if (loggingEvent.getThrowableInformation() == null) {
         writeFormattedException(
-            loggingEvent.getMessage(),
+            loggingEvent.getRenderedMessage(),
             loggingEvent.getLevel(),
             writer);
       } else {
+        Throwable throwable =
+            loggingEvent.getThrowableInformation().getThrowable();
         writeFormattedException(
-            loggingEvent.getMessage(),
-            loggingEvent.getThrowableProxy(),
+            loggingEvent.getRenderedMessage(),
+            throwable,
             loggingEvent.getLevel(),
             writer);
       }
@@ -156,6 +142,15 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
       System.err.println("GEC failed to append: " + e.getMessage());
       e.printStackTrace();
     }
+  }
+
+  @Override
+  public void close() {
+  }
+
+  @Override
+  public boolean requiresLayout() {
+    return false;
   }
 
   /**
@@ -189,7 +184,7 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
       throws IOException {
     JsonGenerator generator = new JsonFactory().createJsonGenerator(out);
 
-    String backtrace = GecAppender.getStackTrace(new Throwable());
+    String backtrace = ExceptionUtils.getFullStackTrace(new Exception());
     String[] lines = backtrace.split("\n");
     StringBuilder builder = new StringBuilder();
     for (String line : lines) {
@@ -228,29 +223,12 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
   void writeFormattedException(final String message,
                                final Throwable throwable,
                                final Level level,
-                               final Writer out) throws IOException {
-    this.writeFormattedException(message,
-        new ThrowableProxy(throwable), level, out);
-  }
-
-  /**
-   * Writes a formatted exception to the given writer.
-   *
-   * @param message   the log message
-   * @param throwableProxy the exception
-   * @param level     the error level
-   * @param out       the destination
-   * @throws IOException if there are IO errors in the destination
-   */
-  private void writeFormattedException(final String message,
-                                       final IThrowableProxy throwableProxy,
-                                       final Level level,
-                                       final Writer out)
+                               final Writer out)
       throws IOException {
     JsonGenerator generator = new JsonFactory().createJsonGenerator(out);
 
-    IThrowableProxy rootThrowable = throwableProxy;
-    while (this.passthroughExceptions.contains(rootThrowable.getClassName())
+    Throwable rootThrowable = throwable;
+    while (this.passthroughExceptions.contains(rootThrowable.getClass())
         && rootThrowable.getCause() != null) {
       rootThrowable = rootThrowable.getCause();
     }
@@ -259,44 +237,17 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
     generator.writeStringField("project", this.project);
     generator.writeStringField("environment", this.environment);
     generator.writeStringField("serverName", this.serverName);
-    // FIXME this was 'throwable'
-    generator.writeStringField("backtrace", getStackTrace(rootThrowable));
+    generator.writeStringField("backtrace",
+        ExceptionUtils.getStackTrace(throwable));
     generator.writeStringField("message", rootThrowable.getMessage());
     generator.writeStringField("logMessage", message);
-    generator.writeStringField("type", rootThrowable.getClassName());
+    generator.writeStringField("type", rootThrowable.getClass().getName());
     if (level != Level.ERROR) {
       generator.writeStringField("errorLevel", level.toString());
     }
     writeContext(generator);
     generator.writeEndObject();
     generator.close();
-  }
-
-
-  /**
-   * Renders a stacktrace.
-   * @param throwableProxy an IThrowableProxy
-   * @return a string rendering of the stack trace
-   */
-  protected static String getStackTrace(final IThrowableProxy throwableProxy) {
-    StringBuilder builder = new StringBuilder();
-    for (StackTraceElementProxy step
-        : throwableProxy.getStackTraceElementProxyArray()) {
-      String string = step.toString();
-      builder.append(CoreConstants.TAB).append(string);
-      ThrowableProxyUtil.subjoinPackagingData(builder, step);
-      builder.append(CoreConstants.LINE_SEPARATOR);
-    }
-    return builder.toString();
-  }
-
-  /**
-   * Renders a stacktrace.
-   * @param t a throwable
-   * @return a string rendering of the stack trace
-   */
-  protected static String getStackTrace(final Throwable t) {
-    return GecAppender.getStackTrace(new ThrowableProxy(t));
   }
 
   /**
@@ -342,7 +293,7 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
    */
   public void addPassthroughExceptionClass(
       final Class<? extends Throwable> exceptionClass) {
-    this.passthroughExceptions.add(exceptionClass.getCanonicalName());
+    this.passthroughExceptions.add(exceptionClass);
   }
 
 
@@ -355,7 +306,12 @@ public final class GecAppender extends AppenderBase<ILoggingEvent> {
    */
   @SuppressWarnings("unchecked")
   public boolean addPassthroughExceptionClass(final String name) {
-    addPassthroughExceptionClass(name);
+    try {
+      addPassthroughExceptionClass(
+          (Class<? extends Throwable>) Class.forName(name));
+    } catch (ClassNotFoundException ex) {
+      return false;
+    }
     return true;
   }
